Update test-list.txt
Hello James,  I started listening/sawing to the Starter Kits l back in Nov/Dec 2019 on the run and mostly on my mobile, but not hands on, nor practice.  I immediately saw the value, and realized your method was what I needed to immerse myself, with my circumstances, to learn something that has been my passion.  I been trying to learn by myself when I see areas I don't know or need clarity, and it has been time-consuming with little results.  So, my aha is that I have to learn to walk before running, and needed to take the little time I have on hands on and practice, practice, practice.  Possible, this will require no to be on the bleeding age of information related to the project, but will provide me solid foundations to build.  - Because of life, I put time, but randomly, to work on learning the under the hood part of programming. Today, I came to review all the basics, and started finding answers.  The question I am asking myself is: How can I benefit form this project and keep building? A possible answer I see would be: Prioritize quality vs quantity by taking and reviewing notes to embrace concepts, and practicing.  Hope this helps, and all comments are welcome.